chore(migration-graph): refactor tests
- Move ember discovery logic to EmberAppProjectGraph.
- Moved tests for migration-graph to corresponding package.
- Removed filterByPackageName logic for now. Will be an another PR.
  - Removed unused deps.
- `migration-graph.test.ts` is focuses on SourceType detection. As that's all it provides now.
- Added API for `fixturify-projects` in `test-support` with `getLibraryProjectFixture()` API.
