+++ packages/migration-graph-ember/test/entities/ember-addon-project-graph.test.ts
@@ -1,0 +1,30 @@
+import { describe, expect, test } from 'vitest';
+import { getEmberProjectFixture } from '@rehearsal/test-support';
+import { EmberAddonProjectGraph } from '../../src/entities/ember-addon-project-graph';
+import type { GraphNode, ModuleNode, UniqueNode } from '@rehearsal/migration-graph-shared';
+
+function flatten(arr: GraphNode<UniqueNode>[]): string[] {
+  return Array.from(arr).map((n) => n.content.key);
+}
+
+function filter(arr: GraphNode<ModuleNode>[]): GraphNode<ModuleNode>[] {
+  return Array.from(arr).filter((n) => {
+    return !n.content.synthetic;
+  });
+}
+
+describe('Unit | EmberAddonProjectGraph', () => {
+  test('addon', async () => {
+    const project = await getEmberProjectFixture('addon');
+
+    const projectGraph = new EmberAddonProjectGraph(project.baseDir);
+    projectGraph.discover();
+
+    const orderedPackages = projectGraph.graph.topSort();
+
+    expect(flatten(orderedPackages)).toStrictEqual(['addon-template']);
+    expect(
+      flatten(filter(orderedPackages[0].content.pkg.getModuleGraph().topSort()))
+    ).toStrictEqual(['addon/components/greet.js']);
+  });
+});
